# src/utils/swagger.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Backend-G7",
    "description": "Students Microservice - Backend API in Node.js with PostgreSQL.",
    "contact": {
      "email": "brunoramos013@gmail.com"
    },
    "version": "v0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "tags": [
    {
      "name": "Student",
      "description": "Get Student Information"
    },
    {
      "name": "Students Management",
      "description": "Manage Students In The System"
    }
  ],
  "paths": {
    "/Students": {
      "post": {
        "tags": [
          "Students Management"
        ],
        "summary": "Create Student",
        "description": " Create a New Student In The System.",
        "operationId": "createStudent",
        "requestBody": {
          "description": "Student Object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Student Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "List Students",
        "description": "Get a List Of Students In The System.",
        "operationId": "listStudents",
        "responses": {
          "200": {
            "description": "Students Listed Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Students/{id}": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "Get Student By ID",
        "description": "List a Student By ID In The System.",
        "operationId": "getStudentByID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID That Needs To Be Fetched.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Student Listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "404": {
            "description": "Student Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Students Management"
        ],
        "summary": "Update Student",
        "description": "Update an existent Student in the System",
        "operationId": "updateStudent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of Student That Needs To Be Updated",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Update An Existent Student In The System",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student Updated"
          },
          "404": {
            "description": "Student Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Students Management"
        ],
        "summary": "Delete Student",
        "description": "Delete An Existent Student In The System.",
        "operationId": "deleteStudent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Insert The ID Of The Student That Needs To Be Deleted",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Student Deleted"
          },
          "404": {
            "description": "Student Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "Students Management"
        ],
        "summary": "Update Student",
        "description": "Update an existent Student in the System",
        "operationId": "updateStudent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Update An Existent Student In The System",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student Updated"
          },
          "404": {
            "description": "Student Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Student": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "required": true
          },
          "createdTimestamp": {
            "type": "number",
            "required": true
          },
          "enabled": {
            "type": "boolean",
            "required": true
          },
          "emailVerified": {
            "type": "boolean",
            "required": true
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "graduation": {
            "type": "string"
          },
          "curriculum": {
            "type": "object",
            "properties": {
              "schooling": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "graduation": {
                      "type": "string"
                    },
                    "institution": {
                      "type": "string"
                    },
                    "conclusion": {
                      "type": "number"
                    }
                  }
                }
              },
              "professional_experience": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "position": {
                      "type": "string"
                    },
                    "contractor": {
                      "type": "string"
                    },
                    "contractTime": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "contractors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "taxpayerIdNum": {
                  "type": "number"
                },
                "name": {
                  "type": "string"
                }
              }
            }
          },
          "requiredActions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notBefore": {
            "type": "number"
          },
          "access": {
            "type": "object",
            "properties": {
              "manageGroupMembership": {
                "type": "boolean"
              },
              "view": {
                "type": "boolean"
              },
              "mapRoles": {
                "type": "boolean"
              },
              "impersonate": {
                "type": "boolean"
              },
              "manage": {
                "type": "boolean"
              }
            }
          }
        }
      }
    }
  }
}